requirements.txt
Flask
Flask-Cors
pandas
google-generativeai
nltk
python-dotenv
faiss-cpu
openpyxl
waitress
